Fix VP9 freezes due to wrong marker bit setting (#1620)
* Fix VP9 freezes due to wrong marker bit setting

Fixes 1616

Even though the spec says:

"
This bit MUST be set to one for the final packet of
the highest spatial-layer frame (the final packet of the picture)
"

Since the receiver endpoint does not know which is the highest spatial-layer
we are sending, marker flag must be set for every final packet of frame.

Tests demonstate this reduces freezes massively.