Worker: enhance SeqManager tests (#1680)
Calling `REQUIRE()` inside a global `validate()` function makes any error in the test point
to the same line. The aim of this PR is that the error points to the exact line in the failing test.

We cannot make `validate()` a marco because the function is a template.

New `validate()` returns a pair of numbers which are later compared in each call. If a test fails
the exact failing test and line are now reported.